+++ spec/osm/badge_spec.rb
@@ -95,9 +95,19 @@ describe "Badge" do
 
   it "Get total requirements gained for a member" do
     data = Osm::Badge::Data.new(
-      :requirements => {'a_1' => 'x', 'a_2' => 'a', 'b_1' => 'yes', 'b_2' => '2000-01-02'}
+      :requirements => {
+        'a_1' => 'a',
+        'a_2' => 'yes',
+        'a_3' => '2000-01-02',
+        'a_4' => 1,
+        'b_1' => 'x',
+        'b_2' => 'xYES',
+        'b_3' => '',
+        'b_4' => nil,
+        'b_5' => 0,
+      }
     )
-    data.total_gained.should == 3
+    data.total_gained.should == 4
   end
 
   it "Get total requirements met in each section for a member" do
@@ -108,37 +118,140 @@ describe "Badge" do
         Osm::Badge::Requirement.new(:field => 'a_2'),
         Osm::Badge::Requirement.new(:field => 'b_1'),
         Osm::Badge::Requirement.new(:field => 'b_2'),
+        Osm::Badge::Requirement.new(:field => 'y_1'),
+        Osm::Badge::Requirement.new(:field => 'y_2'),
       ]
     )
     data = Osm::Badge::Data.new(
       :badge => badge,
-      :requirements => {'a_1' => 'x', 'a_2' => '', 'b_1' => 'yes', 'b_2' => '2000-01-02'}
+      :requirements => {
+        'a_1' => 'x',
+        'a_2' => '',
+        'b_1' => 'yes',
+        'b_2' => '2000-01-02',
+        'y_1' => 1,
+      }
     )
-    data.gained_in_sections.should == {'a' => 0, 'b' => 2}
+    data.gained_in_sections.should == {'a' => 0, 'b' => 2, 'y' => 1}
   end
 
   it "Get number of sections met for a member" do
     badge = Osm::Badge.new(
-      :needed_from_section => {'a' => 1, 'b' => 2},
+      :needed_from_section => {'a' => 1, 'b' => 2, 'c' => 1},
       :requirements => [
         Osm::Badge::Requirement.new(:field => 'a_1'),
         Osm::Badge::Requirement.new(:field => 'a_2'),
         Osm::Badge::Requirement.new(:field => 'b_1'),
         Osm::Badge::Requirement.new(:field => 'b_2'),
+        Osm::Badge::Requirement.new(:field => 'c_1'),
       ]
     )
     data = Osm::Badge::Data.new(
       :badge => badge,
-      :requirements => {'a_1' => 'x', 'a_2' => '', 'b_1' => 'yes', 'b_2' => '2000-01-02'}
+      :requirements => {'a_1' => 'x', 'a_2' => '', 'b_1' => 'yes', 'b_2' => '2000-01-02', 'c_1' => 'yes'}
     )
-    data.sections_gained.should == 1
+    data.sections_gained.should == 2
   end
 
-  it "Works out if the badge is due" do
-    Osm::Badge::Data.new(:completed => 0, :awarded => 0).due?.should be_false
-    Osm::Badge::Data.new(:completed => 1, :awarded => 0).due?.should be_true
-    Osm::Badge::Data.new(:completed => 2, :awarded => 2).due?.should be_false
-    Osm::Badge::Data.new(:completed => 2, :awarded => 1).due?.should be_true
+  describe "Works out if the badge has been earnt" do
+    it "Staged" do
+      badge = Osm::StagedBadge.new(:osm_key => 'not_hikes_or_nights')
+      data = Osm::Badge::Data.new(:awarded => 2, :badge => badge)
+
+      data.stub(:earnt) { 1 }
+      data.earnt?.should be_false
+
+      data.stub(:earnt) { 2 }
+      data.earnt?.should be_false
+
+      data.stub(:earnt) { 3 }
+      data.earnt?.should be_true
+    end
+
+    it "Non staged" do
+      badge = Osm::ActivityBadge.new()
+      data = Osm::Badge::Data.new(:completed => 1, :awarded => 1, :badge => badge)
+      data.earnt?.should be_false
+
+      badge = Osm::ActivityBadge.new()
+      data = Osm::Badge::Data.new(:completed => 1, :awarded => 0, :badge => badge)
+      data.earnt?.should be_true
+
+
+      badge = Osm::ActivityBadge.new(:total_needed => 0, :sections_needed => 2, :needed_from_section => {'a' => 2, 'b' => 1})
+      data = Osm::Badge::Data.new(:requirements => {'a_01'=>'y', 'a_02'=>'y', 'b_01' => 'y'}, :completed => 0, :awarded => 0, :badge => badge)
+      data.earnt?.should be_true
+
+      badge = Osm::ActivityBadge.new(:total_needed => 0, :sections_needed => 2, :needed_from_section => {'a' => 2, 'b' => 1})
+      data = Osm::Badge::Data.new(:requirements => {'a_01'=>'y', 'a_02'=>'y', 'b_01' => 'x'}, :completed => 0, :awarded => 0, :badge => badge)
+      data.earnt?.should be_false
+
+      badge = Osm::ActivityBadge.new(:total_needed => 3, :sections_needed => 0, :needed_from_section => {'a' => 2, 'b' => 1})
+      data = Osm::Badge::Data.new(:requirements => {'a_01'=>'y', 'a_02'=>'y', 'b_01' => 'y'}, :completed => 0, :awarded => 0, :badge => badge)
+      data.earnt?.should be_true
+
+      badge = Osm::ActivityBadge.new(:total_needed => 3, :sections_needed => 0, :needed_from_section => {'a' => 2, 'b' => 1})
+      data = Osm::Badge::Data.new(:requirements => {'a_01'=>'y', 'a_02'=>'x', 'b_01' => 'y'}, :completed => 0, :awarded => 0, :badge => badge)
+      data.earnt?.should be_false
+
+      badge = Osm::ActivityBadge.new(:total_needed => 3, :sections_needed => 2, :needed_from_section => {'a' => 2, 'b' => 1})
+      data = Osm::Badge::Data.new(:requirements => {'a_01'=>'y', 'a_02'=>'y', 'b_01' => 'y'}, :completed => 0, :awarded => 0, :badge => badge)
+      data.earnt?.should be_true
+
+      badge = Osm::ActivityBadge.new(:total_needed => 1, :sections_needed => 1, :needed_from_section => {'a' => 2, 'b' => 1})
+      data = Osm::Badge::Data.new(:requirements => {'a_01'=>'y', 'a_02'=>'y', 'b_01' => 'y'}, :completed => 0, :awarded => 0, :badge => badge)
+      data.earnt?.should be_true
+    end
+  end
+
+  describe "Works out what level of a badge has been earnt" do
+    it "Staged" do
+      badge = Osm::StagedBadge.new(:osm_key => 'not_hikes_or_nights', :needed_from_section => {'a'=>1,'b'=>1,'c'=>1,'d'=>2,'e'=>2})
+
+      data = Osm::Badge::Data.new(:requirements=>{'a_01'=>'','b_01'=>'','c_01'=>'','d_01'=>'','d_02'=>'','e_01'=>'','e_02'=>''}, :badge=>badge)
+      data.earnt.should == 0
+
+      data = Osm::Badge::Data.new(:requirements=>{'a_01'=>'y','b_01'=>'','c_01'=>'','d_01'=>'','d_02'=>'','e_01'=>'','e_02'=>''}, :badge=>badge)
+      data.earnt.should == 1
+
+      data = Osm::Badge::Data.new(:requirements=>{'a_01'=>'y','b_01'=>'y','c_01'=>'','d_01'=>'y','d_02'=>'','e_01'=>'','e_02'=>''}, :badge=>badge)
+      data.earnt.should == 2
+
+      data = Osm::Badge::Data.new(:requirements=>{'a_01'=>'y','b_01'=>'y','c_01'=>'','d_01'=>'y','d_02'=>'y','e_01'=>'','e_02'=>''}, :badge=>badge)
+      data.earnt.should == 4
+    end
+
+    it "Nights away" do
+      badge = Osm::StagedBadge.new(:osm_key => 'nightsaway')
+
+      Osm::Badge::Data.new(:requirements => {'y_01'=>3}, :badge => badge).earnt.should == 1
+      Osm::Badge::Data.new(:requirements => {'y_01'=>5}, :badge => badge).earnt.should == 5
+      Osm::Badge::Data.new(:requirements => {'y_01'=>6}, :badge => badge).earnt.should == 5
+      Osm::Badge::Data.new(:requirements => {'y_01'=>199}, :badge => badge).earnt.should == 175
+      Osm::Badge::Data.new(:requirements => {'y_01'=>200}, :badge => badge).earnt.should == 200
+      Osm::Badge::Data.new(:requirements => {'y_01'=>999}, :badge => badge).earnt.should == 200
+    end
+
+    it "Hikes away" do
+      badge = Osm::StagedBadge.new(:osm_key => 'hikes')
+
+      Osm::Badge::Data.new(:requirements => {'y_01'=>3}, :badge => badge).earnt.should == 1
+      Osm::Badge::Data.new(:requirements => {'y_01'=>5}, :badge => badge).earnt.should == 5
+      Osm::Badge::Data.new(:requirements => {'y_01'=>6}, :badge => badge).earnt.should == 5
+      Osm::Badge::Data.new(:requirements => {'y_01'=>49}, :badge => badge).earnt.should == 35
+      Osm::Badge::Data.new(:requirements => {'y_01'=>50}, :badge => badge).earnt.should == 50
+      Osm::Badge::Data.new(:requirements => {'y_01'=>999}, :badge => badge).earnt.should == 50
+    end
+
+    it "Non staged" do
+      data = Osm::Badge::Data.new(:badge => Osm::ActivityBadge.new)
+
+      data.stub(:earnt?) { true }
+      data.earnt.should == 1
+
+      data.stub(:earnt?) { false }
+      data.earnt.should == 0
+    end
   end
 
   it "Works out if the badge has been started" do
@@ -285,7 +398,7 @@ describe "Badge" do
       end
 
       it "Core" do
-        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=getInitialBadges&type=core&sectionid=1&section=beavers&termid=2", :body => @data)
+        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=getInitialBadges&type=core&sectionid=1&section=beavers&termid=2", :body => @data, :content_type => 'application/json')
         Osm::Term.stub(:get_current_term_for_section){ Osm::Term.new(:id => 2) }
 
         badges = Osm::CoreBadge.get_badges_for_section(@api, Osm::Section.new(:id => 1, :type => :beavers))
@@ -307,7 +420,7 @@ describe "Badge" do
       end
 
       it "Challenge" do
-        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=getInitialBadges&type=challenge&sectionid=1&section=beavers&termid=2", :body => @data)
+        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=getInitialBadges&type=challenge&sectionid=1&section=beavers&termid=2", :body => @data, :content_type => 'application/json')
         Osm::Term.stub(:get_current_term_for_section){ Osm::Term.new(:id => 2) }
 
         badges = Osm::ChallengeBadge.get_badges_for_section(@api, Osm::Section.new(:id => 1, :type => :beavers))
@@ -329,7 +442,7 @@ describe "Badge" do
       end
 
       it "Staged" do
-        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=getInitialBadges&type=staged&sectionid=1&section=beavers&termid=2", :body => @data)
+        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=getInitialBadges&type=staged&sectionid=1&section=beavers&termid=2", :body => @data, :content_type => 'application/json')
         Osm::Term.stub(:get_current_term_for_section){ Osm::Term.new(:id => 2) }
 
         badges = Osm::StagedBadge.get_badges_for_section(@api, Osm::Section.new(:id => 1, :type => :beavers))
@@ -351,7 +464,7 @@ describe "Badge" do
       end
 
       it "Activity" do
-        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=getInitialBadges&type=activity&sectionid=1&section=beavers&termid=2", :body => @data)
+        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=getInitialBadges&type=activity&sectionid=1&section=beavers&termid=2", :body => @data, :content_type => 'application/json')
         Osm::Term.stub(:get_current_term_for_section){ Osm::Term.new(:id => 2) }
 
         badges = Osm::ActivityBadge.get_badges_for_section(@api, Osm::Section.new(:id => 1, :type => :beavers))
@@ -373,7 +486,7 @@ describe "Badge" do
       end
 
       it "For a different section type" do
-        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=getInitialBadges&type=activity&sectionid=1&section=cubs&termid=2", :body => @data)
+        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=getInitialBadges&type=activity&sectionid=1&section=cubs&termid=2", :body => @data, :content_type => 'application/json')
         Osm::Term.stub(:get_current_term_for_section){ Osm::Term.new(:id => 2) }
 
         badges = Osm::ActivityBadge.get_badges_for_section(@api, Osm::Section.new(:id => 1, :type => :beavers), :cubs)
@@ -404,7 +517,7 @@ describe "Badge" do
       end
 
       it "Core badge" do
-        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?termid=2&type=core&section=beavers&c=badge&sectionid=1", :body => @data)
+        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?termid=2&type=core&section=beavers&c=badge&sectionid=1", :body => @data, :content_type => 'application/json')
         datas = Osm::CoreBadge.new(:osm_key => 'badge').get_data_for_section(@api, Osm::Section.new(:id => 1, :type => :beavers), 2)
         datas.size.should == 1
         data = datas[0]
@@ -420,7 +533,7 @@ describe "Badge" do
       end
 
       it "Challenge badge" do
-        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?termid=2&type=challenge&section=beavers&c=badge&sectionid=1", :body => @data)
+        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?termid=2&type=challenge&section=beavers&c=badge&sectionid=1", :body => @data, :content_type => 'application/json')
         datas = Osm::ChallengeBadge.new(:osm_key => 'badge').get_data_for_section(@api, Osm::Section.new(:id => 1, :type => :beavers), 2)
         datas.size.should == 1
         data = datas[0]
@@ -436,7 +549,7 @@ describe "Badge" do
       end
 
       it "Staged badge" do
-        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?termid=2&type=staged&section=beavers&c=badge&sectionid=1", :body => @data)
+        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?termid=2&type=staged&section=beavers&c=badge&sectionid=1", :body => @data, :content_type => 'application/json')
         datas = Osm::StagedBadge.new(:osm_key => 'badge').get_data_for_section(@api, Osm::Section.new(:id => 1, :type => :beavers), 2)
         datas.size.should == 1
         data = datas[0]
@@ -452,7 +565,7 @@ describe "Badge" do
       end
 
       it "Activity badge" do
-        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?termid=2&type=activity&section=beavers&c=badge&sectionid=1", :body => @data)
+        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?termid=2&type=activity&section=beavers&c=badge&sectionid=1", :body => @data, :content_type => 'application/json')
         datas = Osm::ActivityBadge.new(:osm_key => 'badge').get_data_for_section(@api, Osm::Section.new(:id => 1, :type => :beavers), 2)
         datas.size.should == 1
         data = datas[0]
@@ -713,7 +826,7 @@ describe "Badge" do
       end
 
       it "Core badge" do
-        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=summary&section=beavers&sectionid=1&termid=2&type=core", :body => @data)
+        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=summary&section=beavers&sectionid=1&termid=2&type=core", :body => @data, :content_type => 'application/json')
         summary = Osm::CoreBadge.get_summary_for_section(@api, Osm::Section.new(:id => 1, :type => :beavers), 2)
         summary.size.should == 1
         summary[0].should == {
@@ -725,7 +838,7 @@ describe "Badge" do
       end
 
       it "Challenge badge" do
-        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=summary&section=beavers&sectionid=1&termid=2&type=challenge", :body => @data)
+        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=summary&section=beavers&sectionid=1&termid=2&type=challenge", :body => @data, :content_type => 'application/json')
         summary = Osm::ChallengeBadge.get_summary_for_section(@api, Osm::Section.new(:id => 1, :type => :beavers), 2)
         summary.size.should == 1
         summary[0].should == {
@@ -737,7 +850,7 @@ describe "Badge" do
       end
 
       it "Staged badge" do
-        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=summary&section=beavers&sectionid=1&termid=2&type=staged", :body => @data)
+        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=summary&section=beavers&sectionid=1&termid=2&type=staged", :body => @data, :content_type => 'application/json')
         summary = Osm::StagedBadge.get_summary_for_section(@api, Osm::Section.new(:id => 1, :type => :beavers), 2)
         summary.size.should == 1
         summary[0].should == {
@@ -749,7 +862,7 @@ describe "Badge" do
       end
 
       it "Activity badge" do
-        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=summary&section=beavers&sectionid=1&termid=2&type=activity", :body => @data)
+        FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=summary&section=beavers&sectionid=1&termid=2&type=activity", :body => @data, :content_type => 'application/json')
         summary = Osm::ActivityBadge.get_summary_for_section(@api, Osm::Section.new(:id => 1, :type => :beavers), 2)
         summary.size.should == 1
         summary[0].should == {